Rename views folder to viewModules
