+++ src/apps/collectionsUi/viewModules/about/translations.json
