+++ src/apps/collectionsUi/viewModules/home/Async.js
@@ -1,4 +1,4 @@
-import { createAsyncView } from 'modules/bootstrap/higherOrderComponents'
+import { createAsyncView } from 'coreModules/bootstrap/higherOrderComponents'
 
 export default createAsyncView({
   name: 'home',
