+++ src/apps/collectionsUi/viewModules/home/translations.json
