+++ src/apps/collectionsUi/viewModules/index.js
