+++ src/apps/collectionsUi/viewModules/login/translations.json
