+++ src/apps/collectionsUi/viewModules/registerMammal/Async.js
@@ -1,0 +1,14 @@
+import { createAsyncView } from 'coreModules/bootstrap/higherOrderComponents'
+
+export default createAsyncView({
+  modules: () => {
+    return [
+      import('coreModules/form'),
+      import('domainModules/collectionMammals'),
+    ]
+  },
+  name: 'RegisterMammal',
+  view: () => {
+    return import('./index.js')
+  },
+})
