+++ src/apps/collectionsUi/viewModules/registerMammal/index.js
