+++ src/apps/collectionsUi/viewModules/registerMammal/translations.json
