+++ src/apps/collectionsUi/viewModules/settings/index.js
