+++ src/apps/collectionsUi/viewModules/settings/translations.json
