--- src/apps/collectionsUi/views/registerMammal/Async.js
@@ -1,11 +1,0 @@
-import { createAsyncView } from 'modules/bootstrap/higherOrderComponents'
-
-export default createAsyncView({
-  modules: () => {
-    return [import('modules/form'), import('modules/collectionMammals')]
-  },
-  name: 'RegisterMammal',
-  view: () => {
-    return import('./index.js')
-  },
-})
